// src/clue/components/toggle-buttons.scss.d.ts
// cf. https://mattferderer.com/use-sass-variables-in-typescript-and-javascript

export interface IToggleButtonsStyleExports {
  toggleButtonDashboardColor: string;
  toggleButtonDashboardBackgroundColor: string;
  toggleButtonDashboardHoverBackgroundColor: string;
  toggleButtonDashboardSelectedBackgroundColor: string;
  toggleButtonWorkspaceColor: string;
  toggleButtonWorkspaceBackgroundColor: string;
  toggleButtonWorkspaceHoverBackgroundColor: string;
  toggleButtonWorkspaceSelectedBackgroundColor: string;
}

export const styles: IToggleButtonsStyleExports;

export default styles;